% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBcv.fg.R
\name{CoxBcv.fg}
\alias{CoxBcv.fg}
\title{Fay and Graubard (FG) bias-corrected sandwich variance estimator}
\usage{
CoxBcv.fg(Y, Delta, X, ID)
}
\arguments{
\item{Y}{vector of observed time-to-event data.}

\item{Delta}{vector of censoring indicators.}

\item{X}{matrix of marginal mean covariates with one column for one covariate (design matrix excluding intercept).}

\item{ID}{vector of cluster identifiers.}
}
\value{
\itemize{
  \item coef - estimate of coefficients.
  \item exp(coef) - estimate of hazard ratio.
  \item FG-var - FG bias-corrected sandwich variance estimate of coef.
}
}
\description{
Calculate the Fay and Graubard (FG; 2001) bias-corrected sandwich variance estimator, for marginal Cox analysis of cluster randomized trials, 
proposed by Wang et al. (under review).
}
\references{
Fay, M. P., & Graubard, B. I. (2001). 
Small‐sample adjustments for Wald‐type tests using sandwich estimators. 
Biometrics, 57(4), 1198-1206.

Wang, X., Turner, E. L., & Li, F. 
Improving sandwich variance estimation for marginal Cox analysis of cluster randomized trials. 
Under Review.
}
