% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBcv.kcmr.R
\name{CoxBcv.kcmr}
\alias{CoxBcv.kcmr}
\title{Hybrid KCMR bias-corrected sandwich variance estimator}
\usage{
CoxBcv.kcmr(Y, Delta, X, ID)
}
\arguments{
\item{Y}{vector of observed time-to-event data.}

\item{Delta}{vector of censoring indicators.}

\item{X}{matrix of marginal mean covariates with one column for one covariate (design matrix excluding intercept).}

\item{ID}{vector of cluster identifiers.}
}
\value{
\itemize{
  \item coef - estimate of coefficients.
  \item exp(coef) - estimate of hazard ratio.
  \item KCMR-var - KCMR bias-corrected sandwich variance estimate of coef.
}
}
\description{
Calculate the hybrid KCMR bias-corrected sandwich variance estimator, for marginal Cox analysis of cluster randomized trials, 
proposed by Wang et al. (under review). KC: Kauermann and Carroll (2001); MR: martingale residual.
}
\references{
Kauermann, G., & Carroll, R. J. (2001). 
A note on the efficiency of sandwich covariance matrix estimation. 
Journal of the American Statistical Association, 96(456), 1387-1396.

Wang, X., Turner, E. L., & Li, F. 
Improving sandwich variance estimation for marginal Cox analysis of cluster randomized trials. 
Under Review.
}
