% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBcv.md.R
\name{CoxBcv.md}
\alias{CoxBcv.md}
\title{Mancl and DeRouen (MD) bias-corrected sandwich variance estimator}
\usage{
CoxBcv.md(Y, Delta, X, ID)
}
\arguments{
\item{Y}{vector of observed time-to-event data.}

\item{Delta}{vector of censoring indicators.}

\item{X}{matrix of marginal mean covariates with one column for one covariate (design matrix excluding intercept).}

\item{ID}{vector of cluster identifiers.}
}
\value{
\itemize{
  \item coef - estimate of coefficients.
  \item exp(coef) - estimate of hazard ratio.
  \item MD-var - MD bias-corrected sandwich variance estimate of coef.
}
}
\description{
Calculate the Mancl and DeRouen (MD; 2001) bias-corrected sandwich variance estimator, for marginal Cox analysis of cluster randomized trials, 
proposed by Wang et al. (under review).
}
\references{
Mancl, L. A., & DeRouen, T. A. (2001). 
A covariance estimator for GEE with improved small‐sample properties. 
Biometrics, 57(1), 126-134.

Wang, X., Turner, E. L., & Li, F. 
Improving sandwich variance estimation for marginal Cox analysis of cluster randomized trials. 
Under Review.
}
