% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBcv.mr.R
\name{CoxBcv.mr}
\alias{CoxBcv.mr}
\title{Martingale residual (MR) bias-corrected sandwich variance estimator}
\usage{
CoxBcv.mr(Y, Delta, X, ID)
}
\arguments{
\item{Y}{vector of observed time-to-event data.}

\item{Delta}{vector of censoring indicators.}

\item{X}{matrix of marginal mean covariates with one column for one covariate (design matrix excluding intercept).}

\item{ID}{vector of cluster identifiers.}
}
\value{
\itemize{
  \item coef - estimate of coefficients.
  \item exp(coef) - estimate of hazard ratio.
  \item MR-var - MR bias-corrected sandwich variance estimate of coef.
}
}
\description{
Calculate the martingale residual (MR) bias-corrected sandwich variance estimator, for marginal Cox analysis of cluster randomized trials,
proposed by Wang et al. (under review).
}
\references{
Wang, X., Turner, E. L., & Li, F. 
Improving sandwich variance estimation for marginal Cox analysis of cluster randomized trials. 
Under Review.
}
