% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBcv.rob.R
\name{CoxBcv.rob}
\alias{CoxBcv.rob}
\title{Uncorrected robust sandwich variance estimator}
\usage{
CoxBcv.rob(Y, Delta, X, ID)
}
\arguments{
\item{Y}{vector of observed time-to-event data.}

\item{Delta}{vector of censoring indicators.}

\item{X}{matrix of marginal mean covariates with one column for one covariate (design matrix excluding intercept).}

\item{ID}{vector of cluster identifiers.}
}
\value{
\itemize{
  \item coef - estimate of coefficients.
  \item exp(coef) - estimate of hazard ratio.
  \item ROB-var - uncorrected robust sandwich variance estimate of coef.
}
}
\description{
Calculate the uncorrected robust sandwich variance estimator for marginal Cox analysis of cluster randomized trials (Spiekerman and Lin, 1998).
}
\references{
Spiekerman, C. F., & Lin, D. Y. (1998). 
Marginal regression models for multivariate failure time data. 
Journal of the American Statistical Association, 93(443), 1164-1175.
}
